fix: moved RetriggerClearinghouseProcess to maintenance job from clearinghouse
