+++ src/externalsystems/Clearinghouse.Library/BusinessLogic/ClearinghouseBusinessLogic.cs
@@ -37,7 +37,6 @@ public class ClearinghouseBusinessLogic(
     IPortalRepositories portalRepositories,
     IClearinghouseService clearinghouseService,
     IApplicationChecklistService checklistService,
-    IDateTimeProvider dateTimeProvider,
     IOptions<ClearinghouseSettings> options)
     : IClearinghouseBusinessLogic
 {
@@ -143,38 +142,5 @@ public class ClearinghouseBusinessLogic(
             isInvalid
                 ? [ProcessStepTypeId.MANUAL_TRIGGER_OVERRIDE_CLEARING_HOUSE]
                 : [ProcessStepTypeId.START_SELF_DESCRIPTION_LP]);
-    }
-
-    public async Task CheckEndClearinghouseProcesses(CancellationToken cancellationToken)
-    {
-        var applicationIds = await portalRepositories.GetInstance<IApplicationChecklistRepository>()
-            .GetApplicationsForClearinghouseRetrigger(dateTimeProvider.OffsetNow.AddDays(-_settings.RetriggerEndClearinghouseIntervalInDays))
-            .ToListAsync(cancellationToken)
-            .ConfigureAwait(false);
-
-        if (applicationIds.Count == 0)
-            return;
-
-        await foreach (var context in applicationIds
-                                        .Select(applicationId =>
-                                            checklistService.VerifyChecklistEntryAndProcessSteps(
-                                                applicationId,
-                                                ApplicationChecklistEntryTypeId.CLEARING_HOUSE,
-                                                [ApplicationChecklistEntryStatusId.IN_PROGRESS],
-                                                ProcessStepTypeId.AWAIT_CLEARING_HOUSE_RESPONSE))
-                                        .TasksToAsyncEnumerable().WithCancellation(cancellationToken))
-        {
-            checklistService.FinalizeChecklistEntryAndProcessSteps(
-                context,
-                null,
-                item =>
-                {
-                    item.ApplicationChecklistEntryStatusId = ApplicationChecklistEntryStatusId.TO_DO;
-                    item.Comment = "Reset to retrigger clearinghouse";
-                },
-                [ProcessStepTypeId.START_OVERRIDE_CLEARING_HOUSE]);
-        }
-
-        await portalRepositories.SaveAsync().ConfigureAwait(ConfigureAwaitOptions.None);
     }
 }
