+++ src/maintenance/Maintenance.App/DependencyInjection/BatchProcessingServiceSettings.cs
@@ -23,33 +23,39 @@ using System.ComponentModel.DataAnnotations;
 namespace Org.Eclipse.TractusX.Portal.Backend.Maintenance.App.DependencyInjection;
 
 /// <summary>
-/// Settings for the <see cref="BatchDeleteService"/>
+/// Settings for the <see cref="BatchProcessingService"/>
 /// </summary>
-public class BatchDeleteServiceSettings
+public class BatchProcessingServiceSettings
 {
     /// <summary>
     /// Documents older than this configured value will be deleted
     /// </summary>
     [Required]
-    public int DeleteIntervalInDays { get; set; }
+    public int DeleteDocumentsIntervalInDays { get; set; }
+
+    /// <summary>
+    /// Companies waiting for a clearing house validation step that is older than this configured value will be triggered
+    /// </summary>
+    [Required]
+    public int RetriggerClearinghouseIntervalInDays { get; set; }
 }
 
 /// <summary>
-/// Extensions for the <see cref="BatchDeleteService"/>
+/// Extensions for the <see cref="BatchProcessingService"/>
 /// </summary>
-public static class BatchDeleteServiceExtensions
+public static class BatchProcessingServiceExtensions
 {
     /// <summary>
-    /// Adds the <see cref="BatchDeleteService"/> to the service collection
+    /// Adds the <see cref="BatchProcessingService"/> to the service collection
     /// </summary>
     /// <param name="services">The service collection used for di</param>
     /// <param name="section">The configuration section to get the settings from</param>
     /// <returns>The enhanced service collection</returns>
-    public static IServiceCollection AddBatchDelete(this IServiceCollection services, IConfigurationSection section)
+    public static IServiceCollection AddBatchProcessing(this IServiceCollection services, IConfigurationSection section)
     {
-        services.AddOptions<BatchDeleteServiceSettings>().Bind(section);
+        services.AddOptions<BatchProcessingServiceSettings>().Bind(section);
         services
-            .AddTransient<IBatchDeleteService, BatchDeleteService>();
+            .AddTransient<IBatchProcessingService, BatchProcessingService>();
         return services;
     }
 }
