+++ src/maintenance/Maintenance.App/appsettings.json
@@ -23,20 +23,9 @@
   "ConnectionStrings": {
     "PortalDb": "Server=placeholder;Database=placeholder;Port=5432;User Id=placeholder;Password=placeholder;Ssl Mode=Disable;"
   },
-  "BatchDelete": {
-    "DeleteIntervalInDays": 80
-  },
-  "Clearinghouse": {
-    "Username": "",
-    "Password": "",
-    "ClientId": "",
-    "GrantType": "",
-    "ClientSecret": "",
-    "Scope": "",
-    "TokenAddress": "",
-    "BaseAddress": "",
-    "UseDimWallet": false,
-    "RetriggerEndClearinghouseIntervalInDays": 30
+  "BatchProcessing": {
+    "DeleteDocumentsIntervalInDays": 80,
+    "RetriggerClearinghouseIntervalInDays": 30
   },
   "ProcessIdentity": {
     "ProcessUserId": ""
