+++ tests/externalsystems/Clearinghouse.Library.Tests/ClearinghouseBusinessLogicTests.cs
@@ -53,7 +53,6 @@ public class ClearinghouseBusinessLogicTests
     private readonly ClearinghouseBusinessLogic _logic;
     private readonly IClearinghouseService _clearinghouseService;
     private readonly IApplicationChecklistService _checklistService;
-    private readonly IDateTimeProvider _dateTimeProvider;
     private readonly IApplicationChecklistRepository _applicationChecklistRepository;
 
     public ClearinghouseBusinessLogicTests()
@@ -72,10 +71,7 @@ public class ClearinghouseBusinessLogicTests
         A.CallTo(() => _portalRepositories.GetInstance<IApplicationRepository>()).Returns(_applicationRepository);
         A.CallTo(() => _portalRepositories.GetInstance<IApplicationChecklistRepository>()).Returns(_applicationChecklistRepository);
 
-        _dateTimeProvider = A.Fake<IDateTimeProvider>();
-        A.CallTo(() => _dateTimeProvider.OffsetNow).Returns(DateTimeOffset.UtcNow);
-
-        _logic = new ClearinghouseBusinessLogic(_portalRepositories, _clearinghouseService, _checklistService, _dateTimeProvider, Options.Create(new ClearinghouseSettings
+        _logic = new ClearinghouseBusinessLogic(_portalRepositories, _clearinghouseService, _checklistService, Options.Create(new ClearinghouseSettings
         {
             CallbackUrl = "https://api.com"
         }));
@@ -345,29 +341,6 @@ public class ClearinghouseBusinessLogicTests
 
     #endregion
 
-    #region CheckEndClearinghouseProcesses
-
-    [Fact]
-    public async Task CheckEndClearinghouseProcesses_WithEntry_CreatesProcessStep()
-    {
-        // Arrange
-        var entry = new ApplicationChecklistEntry(IdWithBpn, ApplicationChecklistEntryTypeId.IDENTITY_WALLET, ApplicationChecklistEntryStatusId.TO_DO, DateTimeOffset.UtcNow);
-        A.CallTo(() => _applicationChecklistRepository.GetApplicationsForClearinghouseRetrigger(A<DateTimeOffset>._))
-            .Returns(Enumerable.Repeat(IdWithBpn, 1).ToAsyncEnumerable());
-        SetupForCheckEndClearinghouseProcesses(Enumerable.Repeat(entry, 1));
-
-        // Act
-        await _logic.CheckEndClearinghouseProcesses(CancellationToken.None);
-
-        // Assert
-        A.CallTo(() => _checklistService.FinalizeChecklistEntryAndProcessSteps(A<IApplicationChecklistService.ManualChecklistProcessStepData>._, A<Action<ApplicationChecklistEntry>>._, A<Action<ApplicationChecklistEntry>>._, A<IEnumerable<ProcessStepTypeId>>.That.Matches(x => x.Count(y => y == ProcessStepTypeId.START_OVERRIDE_CLEARING_HOUSE) == 1))).MustHaveHappenedOnceExactly();
-        A.CallTo(() => _portalRepositories.SaveAsync()).MustHaveHappenedOnceExactly();
-        entry.Comment.Should().Be("Reset to retrigger clearinghouse");
-        entry.ApplicationChecklistEntryStatusId.Should().Be(ApplicationChecklistEntryStatusId.TO_DO);
-    }
-
-    #endregion
-
     #region Setup
 
     private void SetupForHandleStartClearingHouse()
@@ -415,40 +388,6 @@ public class ClearinghouseBusinessLogicTests
                 A<IEnumerable<ApplicationChecklistEntryTypeId>?>._,
                 A<IEnumerable<ProcessStepTypeId>?>._))
             .Returns(new IApplicationChecklistService.ManualChecklistProcessStepData(Guid.Empty, new Process(Guid.NewGuid(), ProcessTypeId.APPLICATION_CHECKLIST, Guid.NewGuid()), Guid.Empty, ApplicationChecklistEntryTypeId.CLEARING_HOUSE, ImmutableDictionary<ApplicationChecklistEntryTypeId, (ApplicationChecklistEntryStatusId, string?)>.Empty, new List<ProcessStep<Process, ProcessTypeId, ProcessStepTypeId>>()));
-    }
-
-    private void SetupForCheckEndClearinghouseProcesses(IEnumerable<ApplicationChecklistEntry> applicationChecklistEntries)
-    {
-        A.CallTo(() => _checklistService.FinalizeChecklistEntryAndProcessSteps(A<IApplicationChecklistService.ManualChecklistProcessStepData>._, A<Action<ApplicationChecklistEntry>>._, A<Action<ApplicationChecklistEntry>>._, A<IEnumerable<ProcessStepTypeId>>._))
-            .Invokes((IApplicationChecklistService.ManualChecklistProcessStepData data, Action<ApplicationChecklistEntry>? _, Action<ApplicationChecklistEntry> modifyApplicationChecklistEntry, IEnumerable<ProcessStepTypeId> _) =>
-            {
-                var entry = applicationChecklistEntries.SingleOrDefault(x => x.ApplicationId == data.ApplicationId);
-                if (entry == null)
-                    return;
-
-                entry.DateLastChanged = DateTimeOffset.UtcNow;
-                modifyApplicationChecklistEntry(entry);
-            });
-
-        A.CallTo(() => _checklistService.VerifyChecklistEntryAndProcessSteps(
-                A<Guid>._,
-                ApplicationChecklistEntryTypeId.CLEARING_HOUSE,
-                A<IEnumerable<ApplicationChecklistEntryStatusId>>._,
-                ProcessStepTypeId.AWAIT_CLEARING_HOUSE_RESPONSE,
-                A<IEnumerable<ApplicationChecklistEntryTypeId>?>._,
-                A<IEnumerable<ProcessStepTypeId>?>._))
-            .ReturnsLazily((Guid id,
-                ApplicationChecklistEntryTypeId _,
-                IEnumerable<ApplicationChecklistEntryStatusId> _,
-                ProcessStepTypeId _,
-                IEnumerable<ApplicationChecklistEntryTypeId>? _,
-                IEnumerable<ProcessStepTypeId>? _) => new IApplicationChecklistService.ManualChecklistProcessStepData(
-                id,
-                new Process(Guid.NewGuid(), ProcessTypeId.APPLICATION_CHECKLIST, Guid.NewGuid()),
-                Guid.Empty,
-                ApplicationChecklistEntryTypeId.CLEARING_HOUSE,
-                ImmutableDictionary<ApplicationChecklistEntryTypeId, (ApplicationChecklistEntryStatusId, string?)>.Empty,
-                []));
     }
 
     #endregion
