+++ tests/maintenance/Maintenance.App.Tests/appsettings.IntegrationTests.json
@@ -23,20 +23,9 @@
   "ConnectionStrings": {
     "PortalDb": "Server=placeholder;Database=placeholder;Port=5432;User Id=placeholder;Password=placeholder;Ssl Mode=Disable;"
   },
-  "BatchDelete": {
-    "DeleteIntervalInDays": 5
-  },
-  "Clearinghouse": {
-    "Username": "empty",
-    "Password": "empty",
-    "ClientId": "empty",
-    "GrantType": "empty",
-    "ClientSecret": "empty",
-    "Scope": "openid",
-    "TokenAddress": "https://example.org/",
-    "BaseAddress": "https://validation.dev.dih-cloud.com/api/v1",
-    "UseDimWallet": false,
-    "RetriggerEndClearinghouseIntervalInDays": 30
+  "BatchProcessing": {
+    "DeleteDocumentsIntervalInDays": 5,
+    "RetriggerClearinghouseIntervalInDays": 10
   },
   "ProcessIdentity": {
     "ProcessUserId": ""
